pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wikimaker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn==0.34.0",
    "jinja2==3.1.5",
    "trafilatura==2.0.0",
    "httpx==0.28.1",
    "python-frontmatter==1.1.0",
    "python-dotenv==1.0.1",
    "markdown==3.7",
    "python-multipart==0.0.20",
    "youtube-transcript-api==1.2.4",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.scripts]
wikimaker = "app.cli:main"
